user-manager/userManager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-defender==0.9.7
djangorestframework-simplejwt==5.3.0
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
gunicorn==21.2.0